Clumsily added source code for modified builds
The source code for the modified builds has been added to the /src folder.

Note: This source code may not be exactly the same as the already-compiled exe files because it was hastily re-assembled from documentation due to data loss caused by poor decision making. It *should* be the same though.